Issue: Form fields are not updated while travelling in time with the Elm Time Traveling Debugger
Fix: I added the parameter "value" to the input fields with the corresponding value for the model.

In case of the Tags field of the Article, because that is created on the fly buy joining the list of tags, I needed to add an extra value in the model, called "unparsedTags". Otherwise the script will always remove spaces at the end of typed text.

The fix affect these pages:

- http://localhost:8000/#/register
- http://localhost:8000/#/login
- http://localhost:8000/#/editor
- the comment textarea in the Article page