Update comment
I've seen people ask about the "0.19" part of this.

Code splitting and lazy loading were originally planned for 0.19, but as it turned out, it doesn't make sense to wait that long to ship 0.19, so the new plan is to have them in a later release.

I revised the comment to add some clarification, and to no longer reference a specific release. That was probably a mistake to begin with; release scope is always subject to change until it ships!